requirements.txt
aiosqlite
aiohttp
backports.cached-property
beautifulsoup4
bs4
certifi
contextvars
discord
gputil
google-generativeai
moviepy
numpy
opencv-python
pillow
psutil
py-cpuinfo
pylsd-nova
pynacl
pytz
requests
translate
urllib3
websocket
wavelink
youtube-search-python
yt_dlp

# Local AI requirements
# Comment these if you don't plan to use local language models with kobold

langchain
torch
transformers
